// sidecars/tool/browser/src/generated/StreamStatus.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { StreamState } from "./StreamState";
import type { TargetInfo } from "./TargetInfo";

export type StreamStatus = { state: StreamState, browser_gen: number, target: TargetInfo | null, };