// sidecars/tool/browser/src/generated/TargetsMsg.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TargetId } from "./TargetId";
import type { TargetInfo } from "./TargetInfo";

/**
 * At most `LinkLimits::max_targets` tabs.
 */
export type TargetsMsg = { browser_gen: number, targets: Array<TargetInfo>, followed: TargetId | null, };